requirements.txt
django
daphne
channels
channels-redis
djangorestframework
django-cors-headers
requests
paramiko
reportlab
psutil
ping3
beautifulsoup4
pysnmp
cryptography
pymysql
requests-ntlm
pyyaml
jinja2
pyOpenSSL
sqlparse
asgiref
whitenoise
